Add option to use only app/styles and includePath
Putting this here for discussion as it can help with #74.

This PR adds a new option `onlyIncluded` which allows the user to prune the style tree down to just app/styles and anything in the `includePaths` array. This helps performance because the broccoli-caching-writer no longer has to stat everything in the addon trees (via walk-sync).  The speedup is especially pronounced when using linked modules because those trees can get very large.